Beefing up timing script since it'll be helpful for testing how much of a slowdown is caused by profiling
